Simplify Client Secret tab structure and set defaultValue from API
