--- src/app/(pages)/clients/[client]/components/tabs/Credentials/ClientSecret.tsx
@@ -1,9 +1,0 @@
-import Button from "@/components/Button";
-
-export default function ClientSecret() {
-  return (
-    <div className="max-w-54">
-      <Button>Regenerate Client Secret</Button>
-    </div>
-  );
-}
