--- src/app/(pages)/clients/[client]/components/tabs/Credentials/RegistrationAccessToken.tsx
@@ -1,9 +1,0 @@
-import Button from "@/components/Button";
-
-export default function RegistrationAccessToken() {
-  return (
-    <div className="max-w-54">
-      <Button>Regenerate Registration Access Token</Button>
-    </div>
-  );
-}
